Student View changes, adding toastr, VM, html, and js changes
Updated StudentVM to have SelectedCardIndex and CurrentIndex, updated the GetStudentVM method in the Service to set the SelectedCardIndex to the 1st one. Added toastr to the project. Added a text field to the Student UI that shows the currently view card (index) next to "of <NumberOfFlashCards>". The text field is editable to allow the user to switch to a specific card. The text field has an onchange method. The onchange method first performs validation on the newly user-entered index to ensure that an integer is entered and that the integer is greater than 0 and <= the number of flashcards. If the validation fails, an error toastr message will inform them of the issue. If the validation passes, the onchange will then switch the user's currently being viewed flashcard.