Adding a DifficultyLevel drop-down and Enum, merging add/edit for Card into 1
Added a DifficultyLevel enum to accommodate a new drop-down in the add/edit UI.
Refactored the Add/Edit/Save of a Card --
Merged Add/Edit UI's into 1 - Moved all duped code from the Create and Edit views into a AddEditCard view. So, now the same method is called to get the view and to save the card for both the add and the edit.
Took a few methods out of the controller