Form Validation
Re-formatted the AddAssignment form for clarity.
Created a new view "SubmitForm". When the form is validated the data is sent to the new "SubmitForm" view. This can be changed later to accommodate for the necesities of the other areas.
Implemented basic form validation by ensuring the fields are not left blank. Text in the fields cannot be longer than specified.