Helping Mimi with difficulty level color display
As I mentioned to her, this does not yet look the way we want it. I haven't figured out how to display the div beside the other div holding the rest of the card, but it is bringing in the color with it now. But wanted to get her headed in the right direction.