Enhanced Form Validation and more
Added more contraints to the form validation to check for .c or .cpp file names
Fixed some typos
Remove comments and code used for demostration
Refactored the success view used after the form is submitted. Now the view has the correct URL and provides the correct feedback to the user.