Update Details.cshtml
Changed details page linktext "Back to list" to "Back to study page" and changed the asp-action to redirect to student page : the old link redirected to the database page with all questions. User should not have access to this page, since they would be able to change all the information including global flashcards. They should only be able to edit the ones they create.

The correct page to redirect to should be back to student View.